requirements.txt
black
python-dotenv
fastapi
types-requests
uvicorn
./lib/ink